rename device page to devicedemo page for hid device demo
